requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
dnspython==2.4.2
pymongo==4.6.1
ua-parser==0.18.0
user-agents==2.2.0
geoip2==4.7.0
Flask-Limiter==3.5.0
